apply post spoiler warning for notification item contents
